Backend/requirements-cpu.txt
# Django y REST Framework
Django>=5.2.5
djangorestframework>=3.14.0

# Procesamiento de audio (PyTorch CPU)
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
speechbrain>=0.5.14
librosa>=0.10.0
soundfile>=0.12.1

# Procesamiento de imágenes
Pillow>=10.0.0
matplotlib>=3.7.0

# Ciencia de datos
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
transformers>=4.30.0

# Base de datos
psycopg2-binary>=2.9.3

# Utilidades
python-dotenv>=1.0.0
python-magic>=0.4.27
requests>=2.28.0
python-magic-bin>=0.4.14; platform_system == 'Windows'
reportlab>=4.0.0
